Added Frontend for Doctor Registration/Login
