+++ client/package.json
@@ -13,6 +13,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.0.0",
     "react-scripts": "5.0.0",
+    "sass": "^1.52.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
