+++ client/src/App.jsx
@@ -3,6 +3,7 @@ import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
 import LandingPage from "./Pages/Landingpage";
 import DoctorLogin from "./Pages/DoctorLogin";
+import DoctorRegistration from "./Pages/DoctorRegister";
 import Plasmadonor from "./Pages/Plasmadonor";
 import Hospitalbeds from "./Pages/Hospitalbeds";
 import Oxygensuppliers from "./Pages/Oxygensupply";
@@ -25,7 +26,8 @@ function App() {
                     {/* <Navbar /> */}
                     <Switch>
                         <Route path="/" exact component={LandingPage} />
-                        <Route path="/login/doctor" component={DoctorLogin} />
+                        <Route path="/doctor/login" component={DoctorLogin} />
+                        <Route path="/doctor/register" component={DoctorRegistration} />
                         <Route path="/plasma" exact component={Plasmadonor} />
                         <Route path="/hospitalbeds" exact component={Hospitalbeds} />
                         <Route path="/oxygensupply" exact component={Oxygensuppliers} />
