+++ client/src/Pages/DoctorLogin.scss
@@ -1,0 +1,56 @@
+.login {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100vh;
+    background-color: rgb(247, 247, 247);
+
+    .container {
+        max-width: 500px;
+        background-color: #C9ECFF;
+        border-radius: 10px;
+        padding: 30px 50px;
+    }
+
+    form {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        gap: 10px
+    }
+
+    form input {
+        all: unset;
+        padding: 10px;
+        border-radius: 5px;
+        border: 1px solid #ccc;
+        background-color: azure;
+        width: 100%;
+    }
+
+    form select {
+        all: unset;
+        padding: 10px;
+        border-radius: 5px;
+        border: 1px solid #ccc;
+        background-color: azure;
+        width: 100%;
+    }
+
+    .logo {
+        margin-bottom: 20px;
+    }
+
+    .submitBTN {
+        all: unset;
+        padding: 10px;
+        border-radius: 5px;
+        border: 1px solid #ccc;
+        background-color: greenyellow;
+        font-weight: bold;
+        width: 100%;
+        text-align: center;
+    }
+}
